Struct cortex::backend::Backend

pub struct Backend {
    pub connection: Connection,
}

Provides an interface to the Postgres task store

Fields

connection

the Postgres database Connection

Methods

impl Backend

fn from_address(address: &str) -> Backend

Constructs a new Task store representation from a Postgres DB address

fn testdb() -> Backend

Constructs the default Backend struct for testing

fn needs_init(&self) -> bool

Instance methods Checks if the Task store has been initialized, heuristically, by trying to detect if the init service has been added.

fn setup_task_tables(&self) -> Result<()>

Sets up the CorTeX tables and indexes, dropping existing infrastructure when applicable (hard reset)

fn mark_imported(&self, tasks: &Vec<Task>) -> Result<(), Error>

Insert a vector of new Task tasks into the Task store For example, on import, or when a new service is activated on a corpus

fn mark_done(&self, reports: &Vec<TaskReport>) -> Result<(), Error>

Insert a vector of TaskReport reports into the Task store, also marking their tasks as completed with the correct status code.

fn mark_rerun(&self, corpus: &Corpus, service: &Service, severity: Option<String>, category: Option<String>, what: Option<String>) -> Result<(), Error>

Given a complex selector, of a Corpus, Service, and the optional severity, category and what mark all matching tasks to be rerun

fn sync<D: CortexORM + Clone>(&self, d: &D) -> Result<D, Error>

Generic sync method, attempting to obtain the DB record for a given mock Task store datum applicable for any struct implementing the CortexORM trait (for example Corpus, Service, Task)

fn delete<D: CortexORM + Clone>(&self, d: &D) -> Result<(), Error>

Generic delete method, attempting to delete the DB record for a given Task store datum applicable for any struct implementing the CortexORM trait (for example Corpus, Service, Task)

fn add<D: CortexORM + Clone>(&self, d: D) -> Result<D, Error>

Generic addition method, attempting to insert in the DB a Task store datum applicable for any struct implementing the CortexORM trait (for example Corpus, Service, Task)

Note: Overwrites if the entry already existed.

fn fetch_tasks(&self, service: &Service, limit: usize) -> Result<Vec<Task>, Error>

Fetches no more than limit queued tasks for a given Service

fn clear_limbo_tasks(&self) -> Result<(), Error>

Globally resets any "in progress" tasks back to "queued". Particularly useful for dispatcher restarts, when all "in progress" tasks need to be invalidated

fn register_service(&self, service: Service, corpus_path: String) -> Result<(), Error>

Activates an existing service on a given corpus path

fn corpora(&self) -> Vec<Corpus>

Returns a vector of currently available corpora in the Task store

fn progress_report<'report>(&self, c: &Corpus, s: &Service) -> HashMap<String, f64>

Provides a progress report, grouped by severity, for a given Corpus and Service pair

fn task_report<'report>(&self, c: &Corpus, s: &Service, severity: Option<String>, category: Option<String>, what: Option<String>) -> Vec<HashMap<String, String>>

Given a complex selector, of a Corpus, Service, and the optional severity, category and what, Provide a progress report at the chosen granularity