Struct cortex::data::Corpus

pub struct Corpus {
    pub id: Option<i32>,
    pub name: String,
    pub path: String,
    pub complex: bool,
}

A minimal description of a document collection. Defined by a name, path and simple/complex file system setup.

Fields

id

optional id (None for mock / yet-to-be-inserted rows)

name

a human-readable name for this corpus

path

file system path to corpus root (a corpus is held in a single top-level directory)

complex

are we using multiple files to represent a document entry? (if unsure, always use "true")

Methods

impl Corpus

fn select_services<'a>(&'a self, connection: &'a Connection) -> Result<Vec<Service>, Error>

Return a vector of services currently activated on this corpus

fn to_hash(&self) -> HashMap<String, String>

Return a hash representation of the corpus, usually for frontend reports