Struct cortex::data::Service

pub struct Service {
    pub id: Option<i32>,
    pub name: String,
    pub version: f32,
    pub inputformat: String,
    pub outputformat: String,
    pub inputconverter: Option<String>,
    pub complex: bool,
}

A CorTeX processing Service

Fields

id

optional id (None for mock / yet-to-be-inserted rows)

name

a human-readable name for this service

version

a floating-point number to mark the current version (e.g. 0.01)

inputformat

the expected input format for this service (e.g. tex)

outputformat

the produced output format by this service (e.g. html)

inputconverter

prerequisite input conversion service, if any

complex

is this service requiring more than the main textual content of a document? mark "true" if unsure

Methods

impl Service

fn from_name(connection: &Connection, name: String) -> Result<Option<Self>, Error>

Select a service from the Task store via its human-readable name. Requires a postgres Connection.

fn to_hash(&self) -> HashMap<String, String>

Returns a hash representation of the Service, usually for frontend reports