Trait cortex::data::CortexORM

pub trait CortexORM {
    fn select_by_id<'a>(&'a self, connection: &'a Connection) -> Result<Option<Self>, Error> where Self: Sized;
    fn select_by_key<'a>(&'a self, connection: &'a Connection) -> Result<Option<Self>, Error> where Self: Sized;
    fn insert(&self, connection: &Connection) -> Result<(), Error>;
    fn delete(&self, connection: &Connection) -> Result<(), Error>;
    fn from_row(row: Row) -> Self;
    fn get_id(&self) -> Option<i32>;
}

A minimalistic ORM trait for CorTeX data items

Required Methods

fn select_by_id<'a>(&'a self, connection: &'a Connection) -> Result<Option<Self>, Error> where Self: Sized

Select from Task store via the primary id

fn select_by_key<'a>(&'a self, connection: &'a Connection) -> Result<Option<Self>, Error> where Self: Sized

Select from Task store via a struct-specific uniquely identifying key

fn insert(&self, connection: &Connection) -> Result<(), Error>

Inser the row identified by this struct into the Task store (overwrite if present)

fn delete(&self, connection: &Connection) -> Result<(), Error>

Delete the row identified by this struct from the Task store

fn from_row(row: Row) -> Self

Construct a struct from a given Task store row

fn get_id(&self) -> Option<i32>

Obtain the id of the struct, if any