Struct cortex::importer::Importer

pub struct Importer {
    pub corpus: Corpus,
    pub backend: Backend,
    pub cwd: PathBuf,
}

Struct for performing corpus imports into CorTeX

Fields

corpus

a Corpus to be imported, containing all relevant metadata

backend

a Backend on which to persist the import into the Task store

cwd

the current working directory, to resolve relative paths

Methods

impl Importer

fn cwd() -> PathBuf

Convenience method for (recklessly?) obtaining the current working dir

fn unpack(&self) -> Result<(), ()>

Top-level method for unpacking an arxiv-toplogy corpus from its tar-ed form

fn unpack_arxiv_top(&self) -> Result<(), ()>

Unpack the top-level tar files from an arxiv-topology corpus

fn unpack_arxiv_months(&self) -> Result<(), ()>

Unpack the monthly sub-archives of an arxiv-topology corpus, into the CorTeX organization

fn walk_import<'walk>(&self) -> Result<(), Error>

Given a CorTeX-topology corpus, walk the file system and import it into the Task store

fn new_task(&self, entry: String) -> Task

Create a new NoProblem task for the "import" service and the Importer-specified corpus

fn process(&self) -> Result<(), ()>

Top-level import driver, performs an optional unpack, and then an import into the Task store