Struct cortex::manager::Server

pub struct Server {
    pub port: usize,
    pub queue_size: usize,
    pub message_size: usize,
    pub backend: Backend,
    pub backend_address: String,
}

Represents a message queue server, can be used as either dispatcher or a sink

Fields

port

port to listen on

queue_size

the size of the dispatch queue (also the batch size for Task store queue requests)

message_size

size of an individual message chunk sent via zeromq (keep this small to avoid large RAM use, increase to reduce network bandwidth)

backend

a backend struct, in order to reuse the same connection

backend_address

address for the Task store postgres endpoint

Methods

impl Server

fn start_ventilator(&self, services_arc: Arc<Mutex<HashMap<String, Option<Service>>>>, progress_queue_arc: Arc<Mutex<HashMap<i64, TaskProgress>>>, done_queue_arc: Arc<Mutex<Vec<TaskReport>>>, job_limit: Option<usize>) -> Result<(), Error>

Starts a new dispatch Server (ZMQ Ventilator), to serve tasks to processing workers. The ventilator shares state with other manager threads via queues for tasks in progress, as well as a queue for completed tasks pending persisting to disk. A job limit can be provided as a termination condition for the sink server.

fn start_sink(&self, services_arc: Arc<Mutex<HashMap<String, Option<Service>>>>, progress_queue_arc: Arc<Mutex<HashMap<i64, TaskProgress>>>, done_queue_arc: Arc<Mutex<Vec<TaskReport>>>, job_limit: Option<usize>) -> Result<(), Error>

Starts a receiver/sink Server (ZMQ Pull), to accept processing responses. The sink shares state with other manager threads via queues for tasks in progress, as well as a queue for completed tasks pending persisting to disk. A job limit can be provided as a termination condition for the sink server.

fn mark_done_arc(backend: &Backend, reports_arc: &Arc<Mutex<Vec<TaskReport>>>) -> bool

Persists a shared vector of reports to the Task store

fn push_done_queue(reports_arc: &Arc<Mutex<Vec<TaskReport>>>, report: TaskReport)

Adds a task report to a shared report queue