Struct cortex::manager::TaskManager

pub struct TaskManager {
    pub source_port: usize,
    pub result_port: usize,
    pub queue_size: usize,
    pub message_size: usize,
    pub backend_address: String,
}

Manager struct responsible for dispatching and receiving tasks

Fields

source_port

port for requesting/dispatching jobs

result_port

port for responding/receiving results

queue_size

the size of the dispatch queue (also the batch size for Task store queue requests)

message_size

size of an individual message chunk sent via zeromq (keep this small to avoid large RAM use, increase to reduce network bandwidth)

backend_address

address for the Task store postgres endpoint

Methods

impl TaskManager

fn start<'manager>(&'manager self, job_limit: Option<usize>) -> Result<(), Error>

Starts a new manager, spinning of dispatch/sink servers, listening on the specified ports