Struct cortex::models::HistoricalRun

pub struct HistoricalRun {
    pub id: i32,
    pub service_id: i32,
    pub corpus_id: i32,
    pub total: i32,
    pub invalid: i32,
    pub fatal: i32,
    pub error: i32,
    pub warning: i32,
    pub no_problem: i32,
    pub in_progress: i32,
    pub start_time: NaiveDateTime,
    pub end_time: Option<NaiveDateTime>,
    pub owner: String,
    pub description: String,
}

Historical (Corpus, Service) run records

Fields
id: i32

task primary key, auto-incremented by postgresql

service_id: i32

id of the service owning this task

corpus_id: i32

id of the corpus hosting this task

total: i32

total tasks in run

invalid: i32

invalid tasks in run

fatal: i32

fatal results in run

error: i32

error results in run

warning: i32

warning results in run

no_problem: i32

results with no notable problems in run

in_progress: i32

tasks still in progress at end of run

start_time: NaiveDateTime

start timestamp of run

end_time: Option<NaiveDateTime>

end timestamp of run, i.e. timestamp of next run initiation

owner: String

owner who initiated the run

description: String

description of the purpose of this run

Implementations

impl HistoricalRun

pub fn find_by(
corpus: &Corpus,
service: &Service,
connection: &PgConnection
) -> Result<Vec<HistoricalRun>, Error>

Obtain all historical runs for a given (Corpus, Service) pair

pub fn find_current(
corpus: &Corpus,
service: &Service,
connection: &PgConnection
) -> Result<Option<HistoricalRun>, Error>

Obtain a currently ongoing run entry for a (Corpus, Service) pair, if any

pub fn mark_completed(&self, connection: &PgConnection) -> Result<(), Error>

Mark this historical run as completed, by setting end_time to the current time.