Crate cortex

A general purpose processing framework for corpora of scientific documents

Modules

backend
All aggregate operations over the CorTeX PostgresQL store are accessed through the connection of a Backend object.

concerns
Backend models concerns and traits

dispatcher
A ZMQ-based job dispatcher, interfacing between the task Backend and an open set of remote workers

frontend
CorTeX Frontend capabilities - reports, cache, captchas and more

helpers
Helper structures and methods for Task

importer
Import a new corpus into the framework

models
Backend models and traits for the CorTeX "Task store"

reports
Virtual tables/ORM for reports produced by the CorTeX backend

schema
Auto-generated diesel schema for the backend DB

sysinfo
Convenience reports for system info (experimental)

worker
Worker for performing corpus imports, when served as "init" tasks by the CorTeX dispatcher