Crate cortex

Cortex-M processor

This crate provides functionality common to all Cortex-M processors:

Some assembly instructions

Register definition of core peripherals

Exception vector

This crate provides a reset handler that initializes RAM memory and then calls an external start function, which another crate must provide.

All the other exceptions are weakly linked to an abort function, this abort function puts the processor in debug mode (bkpt), and then busy waits.

Lang items like panic_fmt has been stubbed with the abort function

Modules

asm

exception

nvic

peripheral

stubs

Functions

abort
Terminate program

reset
Reset handler