Crate cortex

A general purpose processing framework for corpora of scientific documents

Modules

backend
ORM-like capabilities for high- and mid-level operations on the Task store

data
Data structures and traits for each framework component in the Task store

importer
Import a new corpus into the framework

manager
A ZMQ-based job dispatcher, interfacing between the task Backend and an open set of remote workers

sysinfo
Convenience reports for system info (experimental)

worker
Worker for performing corpus imports, when served as "init" tasks by the CorTeX dispatcher