Module cortex::frontend

CorTeX Frontend capabilities - reports, cache, captchas and more

Modules

cached
Cache-backed logic, currently based on Redis

captcha
Frontend logic responsible for managing ReCaptcha guards

concerns
Common concerns for frontend routes

cors
CORS capabilities for the Rocket frontend

helpers
General purpose auxiliary routines that do not fit the MVC web service paradigm, tending to minor tasks

params
Various parameter data structures for the Rocket frontend routes