Module cortex::frontend::concerns

Common concerns for frontend routes

Constants

UNKNOWN
Placeholder word for unknown filters/fields

Functions

serve_entry
Provide a NamedFile for an entry, redirecting if captcha guard isn't met

serve_entry_preview
Serves an entry as a Template instance to be preview via a client-side asset renderer

serve_report
Prepare a configurable report for a <corpus,server> pair

serve_rerun
Rerun a filtered subset of tasks for a <corpus,service> pair