Module cortex::frontend::helpers

General purpose auxiliary routines that do not fit the MVC web service paradigm, tending to minor tasks

Functions

decorate_uri_encodings
Auto-generates a URI-encoded "foo_uri" entry for each "foo" label associated with a clickable link (for Tera templates)

load_config
Loads the global FrontendConfig from config.json

severity_highlight
Maps a cortex message severity into a bootstrap class for color highlight

uri_escape
TODO: Is this outdated? Maps a regular string into a URI-encoded one

uri_unescape
TODO: Is this outdated? Maps a URI-encoded string into its regular plain text form