Module cortex::models

Backend models and traits for the CorTeX "Task store"

Structs

Corpus
A minimal description of a document collection. Defined by a name, path and simple/complex file system setup.

HistoricalRun
Historical (Corpus, Service) run records

LogError
An error message, as per the LaTeXML convention

LogFatal
A fatal message, as per the LaTeXML convention

LogInfo
An info/debug message, as per the LaTeXML convention

LogInvalid
An invalid message, as per the LaTeXML convention

LogWarning
A warning message, as per the LaTeXML convention

NewCorpus
Insertable Corpus struct

NewHistoricalRun
A new task, to be inserted into CorTeX

NewLogError
A new, insertable, error message

NewLogFatal
A new, insertable, fatal message

NewLogInfo
A new, insertable, info/debug message

NewLogInvalid
A new, insertable, invalid message

NewLogWarning
A new, insertable, warning message

NewService
Insertable struct for Service

NewTask
A new task, to be inserted into CorTeX

NewWorkerMetadata
Metadata collection for workers, updated by the dispatcher upon zmq transactions

RunMetadata
A JSON-friendly data structure, used for the frontend reports

RunMetadataStack
A JSON-friendly data structure, used for vega-lite Stack figures https://vega.github.io/vega-lite/docs/stack.html

Service
A CorTeX processing service

Task
A CorTeX task, for a given corpus-service pair

WorkerMetadata
Metadata collection for workers, updated by the dispatcher upon zmq transactions

Traits

LogRecord
Log actor trait, assumes already Identifiable (for id())